requirements.txt
asgiref==3.2.10
bson==0.5.8
Django==3.1
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
phonenumberslite==8.12.42
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
